% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gDescription}
\alias{gDescription}
\title{Get the description of a NEUMANN mic!}
\usage{
gDescription(model.name = "u87")
}
\arguments{
\item{model.name}{This is the model name of your chosen mic. Please use
one of the model names that results from calling function "gModel".}
}
\value{
character This character vector contains the general description.
of your chosen mic.
}
\description{
This function returns a general description of your chosen mic!
}
\details{
Please tell me something about my favorite mic!
}
\note{
Case in \code{model.name} is ignored.
}
\examples{
gDescription("KU100")
gDescription("tlm49")
}
\keyword{misc}

