% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gMicData}
\alias{gMicData}
\title{Get the technical data of your favorite NEUMANN!}
\usage{
gMicData(model.name = "u87")
}
\arguments{
\item{model.name}{This is the model name of your chosen mic. Please use
one of the model names that results from calling function "gModel". Note
that the technical data is not available for some of the mic models.}
}
\value{
list or data.frame The tables contain the technical data of your
chosen mic.
}
\description{
This function returns a table or a list or tables which contains the
technical data of your chosen NEUMANN microphone.
}
\details{
You are thinking buying a NEUMANN mic, but really want to read
the technical specification before you do that. However, you are
tired of going to the official website and switch between webpages
trying to get the technical data of different models.
Here is your solution!
}
\note{
case in \code{model.name} is ignored
}
\examples{
gMicData("KU100")
gMicData("tlm49")
}
\keyword{misc}

