% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gModel}
\alias{gModel}
\title{Get the list of NEUMANN!}
\usage{
gModel()
}
\value{
character This character vector contains the individual model
names of all current NEUMANN microphone.
}
\description{
This function returns a character vector which holds the model name
of all NEUMANN microphones.
}
\details{
Please tell me all the microphone models that NEUMANN has right now!
}
\examples{
gModel()
}
\keyword{misc}

