% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MicData}
\alias{MicData}
\alias{MicData-package}
\title{My first R package: gameday}
\description{
Get the description and the technical data of a NEUMANN microphone!
}
\details{
Problems come when you want to buy a NEUMANN microphone. What models
does NEUMANN have? What can this mic do for me? Is there a place where
I can get the technical data of my favorite mic? Now use R to solve
your problems!
}

