% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MicData}
\alias{MicData}
\alias{MicData-package}
\title{Package: MicData}
\description{
Get the description and the technical data of a NEUMANN microphone!
}
\details{
Problems come when you want to buy a NEUMANN microphone. What models
does NEUMANN have? What can this mic do for me? Is there a place where
I can get the technical data of my favorite mic? Now use R to solve
your problems!

Three functions are included in this package.
Function "gModel" would give you all the mic models currently available from Neumann.
Function "gDescription" would give you a general description of a specific mic model.
Function "gMicData" would give you the technical data of a specific mic model.

Please refer to the documentation of the each fuction for more information on how to use it!
}

